Statics & the MusicPlayer
This was a pretty interesting lesson and update. Essentially I had to create a GameObject that would hold a music file and have that GameObject persist throughout the game.

That part is fairly easy, as you can just add a piece of code that tells it not to destroy itself on load (.DontDestroyOnLoad). However, when you loop back round to the Start menu again, it duplicates the song.

In order to tackle this, I had to create a static variable that enabled the script to keep an eye on the duplicates. In a nutshell, the static variable checks out to see if any 'instance' of the Music Player exists. If it already exists, it simply stops (or destroys) any new Music Player gameobjects that are created.

If for some reason a Music Player doesn't exist, it'll create one.

Nifty thing to learn and something I'll have to experiment with in my-playground.