Playspace & Ball
Okay, so a bit of simple math wizardry needed in this update, which equated to adding a background image and making sure the camera aligns to it.

I also had to fix something that happened when I moved to Unity 5, where my canvas was no longer adjustable and cropped a great deal of the game space. After speaking to the guys in the GameDev discord, it looks like Unity 5 had added a component for canvas scaling.

Ended up having to switch the UI scaling mode from constant pixel size to scale with screen size. That seemed to do the trick (1280 x 720).

Have also added the first game element - the ball. Next up is adding the physics and the collision stuff.