Paddle & Ball + Launching
Back to the Unity tutorials after a bit of a break - sorry brain! Managed to catch up with what I was doing before and have added some quite extensive comments to explain the logic.

Essentially this was about ensuring that  the ball tracked with the paddle, so as the user moved their mouse the ball would be locked to the paddle and would move with it. This meant implementing a few variables to track that information.

First variable was about (paddleToBallVector) was about capturing the distance between the ball and the paddle and then subtracting it. I could then use that later within the update function to continously check the ball's position against the paddles and then automaticall subtracting the distance so it was always locked.

As for launching the ball, it was all about applying velocity when a mouse down tap was registered. The issue that the tutorial covers is where the ball-to-paddle tracking is in the update function - so even when you launch the ball, it wouldn't display simply because the update was consistenly removing the distance.

So, the trick to solve this was to create a new bool variable that checked if the game had started vs if it hadn't. If the game hadn't started, then the ball-to-paddle element would consistently run. If, however, the mouse was tapped and therefore started the game then it would run the velocity command.

Next up I assume it will be about controlling the boundaries of where the ball can fly (as currently it goes off screen!).