Gravity Update
So, this is a very brief update, but I wanted to log it. Essentially changed some of the Unity project settings (specifically the Physics 2D ones) to alter the game's gravity to -1.

The impact this has is that the ball now has enough velocity to hit the of the game and continually move because of it.